backend(product -search): add suggest endpoint for the use of suggestion while searching, refactor other filter code

+++ backend/microservice-parent/product-search-service/src/main/java/com/promotion/productSearchService/dto/ProductSearchDTO.java
@@ -12,7 +12,7 @@ import java.util.Map;
 @NoArgsConstructor
 public class ProductSearchDTO {
     private String keyword;
-    private Long categoryId;
+    private int categoryId;
     private BigDecimal minPrice;
     private BigDecimal maxPrice;
     private int page = 0;
